Some cleanup in aztables
Moved options and response types to their respective files.
Helper funcs that were used in one place have been removed, their
contents made inline.
Replaced appending to slices with making the correct size and populating
them in place.
