+++ sdk/data/aztables/service_client.go
@@ -12,7 +12,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	generated "github.com/Azure/azure-sdk-for-go/sdk/data/aztables/internal"
 )
 
@@ -92,15 +91,6 @@ func (t *ServiceClient) NewClient(tableName string) *Client {
 	}
 }
 
-// CreateTableOptions contains optional parameters for Client.Create and ServiceClient.CreateTable
-type CreateTableOptions struct {
-	// placeholder for future optional parameters
-}
-
-func (c *CreateTableOptions) toGenerated() *generated.TableClientCreateOptions {
-	return &generated.TableClientCreateOptions{}
-}
-
 // CreateTable creates a table with the specified name. If the service returns a non-successful HTTP status code,
 // the function returns an *azcore.ResponseError type. Specify nil for options if you want to use the default options.
 func (t *ServiceClient) CreateTable(ctx context.Context, name string, options *CreateTableOptions) (CreateTableResponse, error) {
@@ -120,24 +110,6 @@ func (t *ServiceClient) CreateTable(ctx context.Context, name string, options *C
 	}, err
 }
 
-// DeleteTableOptions contains optional parameters for Client.Delete and ServiceClient.DeleteTable
-type DeleteTableOptions struct {
-	// placeholder for future optional parameters
-}
-
-func (c *DeleteTableOptions) toGenerated() *generated.TableClientDeleteOptions {
-	return &generated.TableClientDeleteOptions{}
-}
-
-// DeleteTableResponse contains response fields for ServiceClient.DeleteTable and Client.Delete
-type DeleteTableResponse struct {
-	// placeholder for future optional response fields
-}
-
-func deleteTableResponseFromGen(g generated.TableClientDeleteResponse) DeleteTableResponse {
-	return DeleteTableResponse{}
-}
-
 // DeleteTable deletes a table by name. If the service returns a non-successful HTTP status code, the function returns an *azcore.ResponseError type.
 // Specify nil for options if you want to use the default options.
 func (t *ServiceClient) DeleteTable(ctx context.Context, name string, options *DeleteTableOptions) (DeleteTableResponse, error) {
@@ -145,78 +117,14 @@ func (t *ServiceClient) DeleteTable(ctx context.Context, name string, options *D
 	ctx, endSpan := runtime.StartSpan(ctx, "ServiceClient.DeleteTable", t.client.Tracer(), nil)
 	defer func() { endSpan(err) }()
 
-	resp, err := t.client.Delete(ctx, name, options.toGenerated())
-	if err != nil {
-		return DeleteTableResponse{}, err
-	}
-	return deleteTableResponseFromGen(resp), err
-}
-
-// ListTablesOptions contains optional parameters for ServiceClient.QueryTables
-type ListTablesOptions struct {
-	// OData filter expression.
-	Filter *string
-
-	// Select expression using OData notation. Limits the columns on each record to just those requested, e.g. "$select=PolicyAssignmentId, ResourceId".
-	Select *string
-
-	// Maximum number of records to return.
-	Top *int32
-
-	// NextTableName is the continuation token for the next table to page from
-	NextTableName *string
-}
-
-func (l *ListTablesOptions) toQueryOptions() *generated.QueryOptions {
-	if l == nil {
-		return &generated.QueryOptions{}
-	}
-
-	return &generated.QueryOptions{
-		Filter: l.Filter,
-		Format: to.Ptr(generated.ODataMetadataFormatApplicationJSONODataMinimalmetadata),
-		Select: l.Select,
-		Top:    l.Top,
-	}
-}
-
-// ListTablesResponse contains response fields for ListTablesPager.NextPage
-type ListTablesResponse struct {
-	// NextTableName contains the information returned from the x-ms-continuation-NextTableName header response.
-	NextTableName *string
-
-	// List of tables.
-	Tables []*TableProperties `json:"value,omitempty"`
-}
-
-func fromGeneratedTableQueryResponseEnvelope(g generated.TableClientQueryResponse) ListTablesResponse {
-	var value []*TableProperties
-
-	for _, v := range g.Value {
-		value = append(value, fromGeneratedTableResponseProperties(v))
-	}
-
-	return ListTablesResponse{
-		NextTableName: g.XMSContinuationNextTableName,
-		Tables:        value,
-	}
+	_, err = t.client.Delete(ctx, name, options.toGenerated())
+	return DeleteTableResponse{}, err
 }
 
 // TableProperties contains the properties for a single Table
 type TableProperties struct {
 	// The name of the table.
 	Name *string `json:"TableName,omitempty"`
-}
-
-// Convets a generated TableResponseProperties to a ResponseProperties
-func fromGeneratedTableResponseProperties(g *generated.TableResponseProperties) *TableProperties {
-	if g == nil {
-		return nil
-	}
-
-	return &TableProperties{
-		Name: g.TableName,
-	}
 }
 
 // NewListTablesPager queries the existing tables using the specified ListTablesOptions.
@@ -259,30 +167,21 @@ func (t *ServiceClient) NewListTablesPager(listOptions *ListTablesOptions) *runt
 			if err != nil {
 				return ListTablesResponse{}, err
 			}
-			return fromGeneratedTableQueryResponseEnvelope(resp), nil
+
+			tableProps := make([]*TableProperties, len(resp.Value))
+			for i := range resp.Value {
+				tableProps[i] = &TableProperties{
+					Name: resp.Value[i].TableName,
+				}
+			}
+
+			return ListTablesResponse{
+				NextTableName: resp.XMSContinuationNextTableName,
+				Tables:        tableProps,
+			}, nil
 		},
 		Tracer: t.client.Tracer(),
 	})
-}
-
-// GetStatisticsOptions contains optional parameters for ServiceClient.GetStatistics
-type GetStatisticsOptions struct {
-	// placeholder for future optional parameters
-}
-
-// GetStatisticsResponse contains response fields for Client.GetStatistics
-type GetStatisticsResponse struct {
-	GeoReplication *GeoReplication `xml:"GeoReplication"`
-}
-
-func getStatisticsResponseFromGenerated(g *generated.ServiceClientGetStatisticsResponse) GetStatisticsResponse {
-	return GetStatisticsResponse{
-		GeoReplication: fromGeneratedGeoReplication(g.GeoReplication),
-	}
-}
-
-func (g *GetStatisticsOptions) toGenerated() *generated.ServiceClientGetStatisticsOptions {
-	return &generated.ServiceClientGetStatisticsOptions{}
 }
 
 // GetStatistics retrieves all the statistics for an account with Geo-redundancy established. If the service returns a non-successful
@@ -299,36 +198,9 @@ func (t *ServiceClient) GetStatistics(ctx context.Context, options *GetStatistic
 	if err != nil {
 		return GetStatisticsResponse{}, err
 	}
-	return getStatisticsResponseFromGenerated(&resp), err
-}
-
-// GetPropertiesOptions contains optional parameters for Client.GetProperties
-type GetPropertiesOptions struct {
-	// placeholder for future optional parameters
-}
-
-func (g *GetPropertiesOptions) toGenerated() *generated.ServiceClientGetPropertiesOptions {
-	return &generated.ServiceClientGetPropertiesOptions{}
-}
-
-// GetPropertiesResponse contains response fields for Client.GetProperties
-type GetPropertiesResponse struct {
-	ServiceProperties
-}
-
-func getPropertiesResponseFromGenerated(g *generated.ServiceClientGetPropertiesResponse) GetPropertiesResponse {
-	var cors []*CorsRule
-	for _, c := range g.Cors {
-		cors = append(cors, fromGeneratedCors(c))
-	}
-	return GetPropertiesResponse{
-		ServiceProperties: ServiceProperties{
-			Cors:          cors,
-			HourMetrics:   fromGeneratedMetrics(g.HourMetrics),
-			Logging:       fromGeneratedLogging(g.Logging),
-			MinuteMetrics: fromGeneratedMetrics(g.MinuteMetrics),
-		},
-	}
+	return GetStatisticsResponse{
+		GeoReplication: fromGeneratedGeoReplication(resp.GeoReplication),
+	}, nil
 }
 
 // GetProperties retrieves the properties for an account including the metrics, logging, and cors rules established.
@@ -346,25 +218,21 @@ func (t *ServiceClient) GetProperties(ctx context.Context, options *GetPropertie
 	if err != nil {
 		return GetPropertiesResponse{}, err
 	}
-	return getPropertiesResponseFromGenerated(&resp), err
-}
-
-// SetPropertiesOptions contains optional parameters for Client.SetProperties
-type SetPropertiesOptions struct {
-	// placeholder for future optional parameters
-}
-
-func (s *SetPropertiesOptions) toGenerated() *generated.ServiceClientSetPropertiesOptions {
-	return &generated.ServiceClientSetPropertiesOptions{}
-}
-
-// SetPropertiesResponse contains response fields for Client.SetProperties
-type SetPropertiesResponse struct {
-	// placeholder for future response fields
-}
-
-func setPropertiesResponseFromGenerated(g *generated.ServiceClientSetPropertiesResponse) SetPropertiesResponse {
-	return SetPropertiesResponse{}
+	var cors []*CorsRule
+	if len(resp.Cors) > 0 {
+		cors = make([]*CorsRule, len(resp.Cors))
+		for i := range resp.Cors {
+			cors[i] = fromGeneratedCors(resp.Cors[i])
+		}
+	}
+	return GetPropertiesResponse{
+		ServiceProperties: ServiceProperties{
+			Cors:          cors,
+			HourMetrics:   fromGeneratedMetrics(resp.HourMetrics),
+			Logging:       fromGeneratedLogging(resp.Logging),
+			MinuteMetrics: fromGeneratedMetrics(resp.MinuteMetrics),
+		},
+	}, nil
 }
 
 // SetProperties allows the user to set cors, metrics, and logging rules for the account.
@@ -386,11 +254,8 @@ func (t *ServiceClient) SetProperties(ctx context.Context, properties ServicePro
 	if options == nil {
 		options = &SetPropertiesOptions{}
 	}
-	resp, err := t.service.SetProperties(ctx, *properties.toGenerated(), options.toGenerated())
-	if err != nil {
-		return SetPropertiesResponse{}, err
-	}
-	return setPropertiesResponseFromGenerated(&resp), err
+	_, err = t.service.SetProperties(ctx, *properties.toGenerated(), options.toGenerated())
+	return SetPropertiesResponse{}, err
 }
 
 // GetAccountSASURL is a convenience method for generating a SAS token for the currently pointed at account. This methods returns the full service URL and an error
